Added many new SQL Queries

+++ src/main/kotlin/com/example/modules/Company.kt
@@ -5,7 +5,7 @@ import java.util.Date
 data class Company(
     val id: String,
     val name : String,
-    val current_price : Float
+    val current_price : Double
 )
 
 val companies = mutableListOf<Company>()
